feat: recover blocked connectors (intake, heuristic, vendor specs) + modularize importer
Address connector-import gaps and split the god files:
- Refactor: internal/connectorimport -> importer/summary/catalog.go; CLI ->
  main/manifest/spec/apisguru/catalogwrite/report.go (each focused, <160 LOC).
- Spec intake: prefer JSON spec URL and sanitize Unicode control characters that
  APIs.guru YAML embeds (recovers DocuSign, SendGrid, bunq).
- openapigen: extend high-value coverage vocabulary with security/identity/GRC
  terms so genuinely sensitive connectors clear the proof gate (recovers Conjur,
  NeutrinoAPI), with a regression test guarding against business-object gaming.
- Manifest: fix personio key and add curated vendor specs (Cloudflare, Tailscale)
  for marquee providers absent from APIs.guru.

Catalog 195 -> 202; catalog-check -require-sourcegen-ready and sourcegen-test green.
