+++ internal/connectorcatalog/catalog/collaboration-productivity.yaml
@@ -3929,3 +3929,209 @@ entries:
             - 200
           method: GET
           path: /subaccounts
+  - classifier_output: supported
+    definition:
+      auth:
+        credential_fields:
+          - key: api_key
+            label: API key
+            reference_only: true
+            required: true
+            secret: true
+        model: api_key
+        requires_references: true
+      categories:
+        - saas
+      description: |-
+        The Beta endpoints for the new Email Activity APIs - functionality is subject to change without notice. You may not have access to this Beta endpoint.
+
+        Email Activity offers filtering and search by event type for two days worth of data. There is an optional add-on to store 60 days worth of data. This add-on also gives you access to the ability to download a CSV of the 60 days worth of email event data. The Beta endpoints for the new Email Activity APIs - functionality is subject to change without notice. You may not have access to this Beta endpoint.
+
+        Email Activity offers filtering and search by event type for two days worth of data. There is an optional add-on to store 60 days worth of data. This add-on also gives you access to the ability to download a CSV of the 60 days worth of email event data.
+      display_name: SendGrid
+      id: builtin_catalog-sendgrid
+      resource_families:
+        - coverage:
+            - families:
+                - activity
+              high_value: true
+              id: audit_events
+              support: partial
+              title: Activity
+              type: audit_event
+          default_enabled: true
+          event:
+            kind: sendgrid.activity
+            required_payload_fields:
+              - allowed
+            schema_ref: sendgrid/activity/v1
+            urn_kind: sendgrid_activity
+          event_kind: activity
+          id: activity
+          id_field: allowed
+          label: Activity
+          list_key: result
+          method: GET
+          name_field: allowed
+          pagination:
+            page_size: 100
+            page_size_param: limit
+            type: page
+          path: /access_settings/activity
+          permissions_needed:
+            - Authorization
+          projection:
+            fields:
+              id: allowed
+              name: allowed
+              provider_id: allowed
+            template: audit_event
+          record_selector: $.result[*]
+        - coverage:
+            - families:
+                - api_key
+              high_value: true
+              id: secrets
+              support: partial
+              title: Api Key
+              type: entity_family
+          default_enabled: true
+          event:
+            kind: sendgrid.api_key
+            required_payload_fields:
+              - name
+            schema_ref: sendgrid/api_key/v1
+            urn_kind: sendgrid_api_key
+          event_kind: api_key
+          id: api_key
+          id_field: name
+          label: Api Key
+          list_key: result
+          method: GET
+          name_field: name
+          pagination:
+            page_size: 100
+            page_size_param: limit
+            type: page
+          path: /api_keys
+          permissions_needed:
+            - Authorization
+          projection:
+            fields:
+              id: name
+              name: name
+              secret_id: name
+              secret_name: name
+              secret_type: api_key
+            template: secret
+          record_selector: $.result[*]
+        - coverage:
+            - families:
+                - group
+              high_value: true
+              id: group
+              support: partial
+              title: Group
+              type: entity_family
+          default_enabled: true
+          event:
+            kind: sendgrid.group
+            required_payload_fields:
+              - id
+            schema_ref: sendgrid/group/v1
+            urn_kind: sendgrid_group
+          event_kind: group
+          id: group
+          id_field: id
+          label: Group
+          method: GET
+          name_field: name
+          pagination:
+            type: none
+          path: /asm/groups
+          permissions_needed:
+            - Authorization
+          projection:
+            fields:
+              id: id
+              name: name
+              provider_id: id
+            template: identity_group
+          record_selector: $[*]
+        - coverage:
+            - families:
+                - invalid_email
+              high_value: true
+              id: invalid_email
+              support: partial
+              title: Invalid Email
+              type: entity_family
+          default_enabled: true
+          event:
+            kind: sendgrid.invalid_email
+            required_payload_fields:
+              - email
+            schema_ref: sendgrid/invalid_email/v1
+            urn_kind: sendgrid_invalid_email
+          event_kind: invalid_email
+          id: invalid_email
+          id_field: email
+          label: Invalid Email
+          method: GET
+          name_field: email
+          pagination:
+            limit_param: limit
+            offset_param: offset
+            page_size: 100
+            type: offset
+          path: /suppression/invalid_emails
+          permissions_needed:
+            - Authorization
+          projection:
+            fields:
+              id: email
+              name: email
+              provider_id: email
+            template: identity_user
+          record_selector: $[*]
+      schema_version: cerebro.integration/v1
+      scope_options:
+        - default_enabled: true
+          families:
+            - activity
+          id: activity
+          label: Activity
+        - default_enabled: true
+          families:
+            - api_key
+          id: api_key
+          label: Api Key
+        - default_enabled: true
+          families:
+            - group
+          id: group
+          label: Group
+        - default_enabled: true
+          families:
+            - invalid_email
+          id: invalid_email
+          label: Invalid Email
+      source_id: sendgrid
+      tenant_id: builtin_catalog
+      transport:
+        base_url: http://api.sendgrid.com/v3
+        retry:
+          backoff: exponential
+          max_attempts: 3
+          retry_after_header: Retry-After
+          statuses:
+            - 429
+            - 500
+            - 502
+            - 503
+            - 504
+        verification:
+          expect_status:
+            - 200
+          method: GET
+          path: /access_settings/activity
