+++ internal/connectorcatalog/catalog/security-posture-vulnerability.yaml
@@ -3839,3 +3839,219 @@ entries:
             - 200
           method: GET
           path: /api/patch/v1/export/advisories
+  - classifier_output: supported
+    definition:
+      auth:
+        credential_fields:
+          - key: api_key
+            label: API key
+            reference_only: true
+            required: true
+            secret: true
+        model: api_key
+        requires_references: true
+      categories:
+        - networking
+        - security
+      config_fields:
+        - help: Path parameter inferred from the OpenAPI operation.
+          key: account_id
+          label: Account Id
+          required: true
+      description: Cloudflare CDN, DNS, WAF, and zero-trust security platform.
+      display_name: Cloudflare
+      id: builtin_catalog-cloudflare
+      resource_families:
+        - coverage:
+            - families:
+                - asn
+              high_value: true
+              id: findings
+              support: partial
+              title: Asn
+              type: remediation_state
+          default_enabled: true
+          event:
+            kind: cloudflare.asn
+            required_payload_fields:
+              - code
+            schema_ref: cloudflare/asn/v1
+            urn_kind: cloudflare_asn
+          event_kind: asn
+          id: asn
+          id_field: code
+          label: Asn
+          list_key: errors
+          method: GET
+          name_field: code
+          pagination:
+            type: none
+          path: /accounts/${config.account_id}/botnet_feed/configs/asn
+          permissions_needed:
+            - api_email
+            - api_key
+            - api_token
+          projection:
+            fields:
+              finding_id: code
+              id: code
+              name: code
+              severity: code
+              status: code
+              title: code
+            template: finding
+          record_selector: $.errors[*]
+        - coverage:
+            - families:
+                - event
+              high_value: true
+              id: audit_events
+              support: partial
+              title: Event
+              type: audit_event
+          default_enabled: true
+          event:
+            kind: cloudflare.event
+            required_payload_fields:
+              - code
+            schema_ref: cloudflare/event/v1
+            urn_kind: cloudflare_event
+          event_kind: event
+          id: event
+          id_field: code
+          label: Event
+          list_key: errors
+          method: GET
+          name_field: code
+          pagination:
+            type: none
+          path: /user/load_balancing_analytics/events
+          permissions_needed:
+            - api_email
+            - api_key
+            - api_token
+          projection:
+            fields:
+              id: code
+              name: code
+              provider_id: code
+            template: audit_event
+          record_selector: $.errors[*]
+        - coverage:
+            - families:
+                - account
+              high_value: true
+              id: account
+              support: partial
+              title: Account
+              type: entity_family
+          default_enabled: true
+          event:
+            kind: cloudflare.account
+            required_payload_fields:
+              - code
+            schema_ref: cloudflare/account/v1
+            urn_kind: cloudflare_account
+          event_kind: account
+          id: account
+          id_field: code
+          label: Account
+          list_key: errors
+          method: GET
+          name_field: code
+          pagination:
+            inject_first_page: true
+            page_param: page
+            page_size: 100
+            page_size_param: per_page
+            start_page: 1
+            type: page
+          path: /accounts
+          permissions_needed:
+            - api_email
+            - api_key
+          projection:
+            fields:
+              id: code
+              name: code
+              provider_id: code
+            template: identity_user
+          record_selector: $.errors[*]
+        - coverage:
+            - families:
+                - organization
+              high_value: true
+              id: organization
+              support: partial
+              title: Organization
+              type: entity_family
+          default_enabled: true
+          event:
+            kind: cloudflare.organization
+            required_payload_fields:
+              - id
+            schema_ref: cloudflare/organization/v1
+            urn_kind: cloudflare_organization
+          event_kind: organization
+          id: organization
+          id_field: id
+          label: Organization
+          list_key: errors
+          method: GET
+          pagination:
+            cursor_param: page_token
+            page_size: 100
+            page_size_param: page_size
+            type: cursor
+          path: /organizations
+          permissions_needed:
+            - api_email
+            - api_key
+            - api_token
+          projection:
+            fields:
+              id: id
+              provider_id: id
+            template: identity_group
+          record_selector: $.errors[*]
+      schema_version: cerebro.integration/v1
+      scope_options:
+        - default_enabled: true
+          families:
+            - asn
+          id: asn
+          label: Asn
+        - default_enabled: true
+          families:
+            - event
+          id: event
+          label: Event
+        - default_enabled: true
+          families:
+            - account
+          id: account
+          label: Account
+        - default_enabled: true
+          families:
+            - organization
+          id: organization
+          label: Organization
+      source_id: cloudflare
+      tenant_id: builtin_catalog
+      transport:
+        base_url: https://api.cloudflare.com/client/v4
+        retry:
+          backoff: exponential
+          max_attempts: 3
+          retry_after_header: Retry-After
+          statuses:
+            - 429
+            - 500
+            - 502
+            - 503
+            - 504
+        verification:
+          expect_status:
+            - 200
+          method: GET
+          path: /accounts/${config.account_id}/botnet_feed/configs/asn
