+++ internal/connectorcatalog/catalog_test.go
@@ -149,13 +149,13 @@ func TestBuiltinCatalogSeedSummary(t *testing.T) {
 	if len(analysis.Issues) != 0 {
 		t.Fatalf("issues = %#v, want none", analysis.Issues)
 	}
-	if analysis.Summary.Total != 195 {
-		t.Fatalf("summary total = %d, want 195", analysis.Summary.Total)
+	if analysis.Summary.Total != 202 {
+		t.Fatalf("summary total = %d, want 202", analysis.Summary.Total)
 	}
-	if len(analysis.Entries) != 195 {
-		t.Fatalf("entries len = %d, want 195", len(analysis.Entries))
+	if len(analysis.Entries) != 202 {
+		t.Fatalf("entries len = %d, want 202", len(analysis.Entries))
 	}
-	if analysis.Summary.Generateable != 195 {
+	if analysis.Summary.Generateable != 202 {
 		t.Fatalf("summary = %#v, want all entries generateable", analysis.Summary)
 	}
 	if analysis.Summary.NeedsAuthExtension != 0 {
