+++ internal/connectordefinitions/openapigen/highvalue_test.go
@@ -1,0 +1,29 @@
+package openapigen
+
+import "testing"
+
+func TestIsHighValueSecurityVocabulary(t *testing.T) {
+	highValue := []struct{ family, path string }{
+		{"credentials", "/v1/credentials"},
+		{"audit_logs", "/v1/audit/logs"},
+		{"sessions", "/api/sessions"},
+		{"certificates", "/v1/certificates"},
+	}
+	for _, c := range highValue {
+		if !isHighValue("", c.family, c.path) {
+			t.Errorf("isHighValue(%q,%q) = false, want true (security/GRC vocabulary)", c.family, c.path)
+		}
+	}
+	// Generic business objects must stay low-value so the catalog proof gate is
+	// not gamed into accepting connectors with no security/GRC coverage.
+	lowValue := []struct{ family, path string }{
+		{"invoices", "/v1/invoices"},
+		{"payments", "/v1/payments"},
+		{"products", "/v1/products"},
+	}
+	for _, c := range lowValue {
+		if isHighValue("", c.family, c.path) {
+			t.Errorf("isHighValue(%q,%q) = true, want false (generic business object)", c.family, c.path)
+		}
+	}
+}
