+++ tools/connectorimport/catalogwrite.go
@@ -1,0 +1,89 @@
+// Staging and built-in-catalog file writers.
+package main
+
+import (
+	"fmt"
+	"os"
+	"path/filepath"
+	"sort"
+	"strings"
+
+	"github.com/writer/cerebro/internal/connectorimport"
+)
+
+func writeStagingCatalogs(outDir string, outcomes []connectorimport.Outcome) error {
+	byDomain := map[string][]connectorimport.Outcome{}
+	for _, outcome := range outcomes {
+		if !outcome.CatalogReady() {
+			continue
+		}
+		domain := strings.TrimSpace(outcome.Domain)
+		if domain == "" {
+			domain = "uncategorized"
+		}
+		byDomain[domain] = append(byDomain[domain], outcome)
+	}
+	for domain, entries := range byDomain {
+		header := fmt.Sprintf("Candidate connector definitions for %s.\nGenerated by tools/connectorimport; review before promotion.", domain)
+		body, err := connectorimport.RenderCatalogEntries(header, entries)
+		if err != nil {
+			return err
+		}
+		path := filepath.Join(outDir, domain+".yaml")
+		if err := os.WriteFile(path, body, 0o600); err != nil {
+			return fmt.Errorf("write %s: %w", path, err)
+		}
+	}
+	return nil
+}
+
+func appendToCatalog(catalogDir string, outcomes []connectorimport.Outcome) (int, error) {
+	blocks, err := connectorimport.RenderCatalogEntryBlocks(outcomes)
+	if err != nil {
+		return 0, err
+	}
+	byDomain := map[string][]connectorimport.Outcome{}
+	for _, outcome := range outcomes {
+		if outcome.CatalogReady() {
+			byDomain[strings.TrimSpace(outcome.Domain)] = append(byDomain[strings.TrimSpace(outcome.Domain)], outcome)
+		}
+	}
+	appended := 0
+	for domain, entries := range byDomain {
+		path := filepath.Join(catalogDir, domain+".yaml")
+		existing, err := os.ReadFile(path) //nolint:gosec // operator-provided catalog dir for a build-time tool.
+		if err != nil {
+			return appended, fmt.Errorf("read catalog file %s (is -append-catalog domain valid?): %w", path, err)
+		}
+		present := existingSourceIDs(string(existing))
+		var additions strings.Builder
+		sort.SliceStable(entries, func(i, j int) bool { return entries[i].SourceID < entries[j].SourceID })
+		for _, outcome := range entries {
+			if _, ok := present[outcome.SourceID]; ok {
+				continue
+			}
+			additions.WriteString(strings.TrimRight(blocks[outcome.SourceID], "\n"))
+			additions.WriteString("\n")
+			appended++
+		}
+		if additions.Len() == 0 {
+			continue
+		}
+		merged := strings.TrimRight(string(existing), "\n") + "\n" + additions.String()
+		if err := os.WriteFile(path, []byte(merged), 0o600); err != nil {
+			return appended, fmt.Errorf("write catalog file %s: %w", path, err)
+		}
+	}
+	return appended, nil
+}
+
+func existingSourceIDs(body string) map[string]struct{} {
+	ids := map[string]struct{}{}
+	for _, line := range strings.Split(body, "\n") {
+		trimmed := strings.TrimSpace(line)
+		if strings.HasPrefix(trimmed, "source_id:") {
+			ids[strings.TrimSpace(strings.TrimPrefix(trimmed, "source_id:"))] = struct{}{}
+		}
+	}
+	return ids
+}
